requirements.txt
fastapi
uvicorn
sqlalchemy
PyJWT
python-dotenv
cryptography
google-generativeai
openai
pydantic
Flask